Trait r3_support_rp2040::usbstdio::Options

pub trait Options: 'static + Send + Sync {
    fn handle_input(_s: &[u8]) { ... }
    fn product_name() -> &'static str { ... }
    fn should_pause_output() -> bool { ... }
}

The options for the USB serial device configured by configure.

Provided Methods

fn handle_input(_s: &[u8])

Handle incoming data.

This method may be called with interrupts disabled. It’s safe to write bytes to the USB serial device here.

fn product_name() -> &'static str

Get the product name to indicate in the USB device descriptor.

fn should_pause_output() -> bool

Return a flag indicating whether the output data should be withheld from transmission.

If this flag is changed to false, poll must be called to flush the data in the transmission buffer.